Add target filtering for logs (#37)
Adding a target_filter attribute for LogExportTargets struct. This will
help users attach a target to their logs and make sure only a subset of
the logs are going to be exported to the Otel collector

Unrelated to the main feature, there is a make test failure in mocks.rs.
The [clippy::result_large_err] lint warns when the [Err] variant of a
Result<T, E> type is very large (in this case, 176 bytes). This can be
problematic because:
- Memory Usage: Result<T, E> is stored as a tagged union, so it takes up
space equal to the larger of T or E plus some overhead for the
discriminant tag.
- Performance: Large error types can impact performance during:
 Stack allocation/deallocation, Moving/copying the Result around, etc

Why is tonic::Status Large?
The tonic::Status type is large because it contains:
Error code, Error message (String), Error details, Metadata, Source
information. All of this adds up to 176+ bytes.
But since this code is:
- Test/mock code (not production)
- A simple interceptor that rarely fails
- Following the tonic interceptor API signature
The #[allow(clippy::result_large_err)] is reasonable here because:
- We can't easily change the API (it's defined by tonic)
- The performance impact is minimal in tests
- Boxing the error would add unnecessary heap allocation for test code